refactor: extract address table, asset locks, and single-key view from wallets_screen
Continue the wallets_screen/mod.rs refactoring started in #539 (dialogs
extraction). Extract three more focused submodules:

- address_table.rs: SortColumn/SortOrder enums, AddressData struct,
  sorting logic, categorize_path, and the full address table renderer
- asset_locks.rs: asset lock table rendering and fund dialog triggers
- single_key_view.rs: single-key wallet detail view with UTXO display

Reduces mod.rs from 2662 to 1947 lines (-27%).
